// skills-server/src/core/helpers/api-text.enum.ts
export enum ApiTextEnum {
  AUTH = 'Auth',
  LOGIN = 'Login',
  USERS = 'Users',
  CONTACTS = 'Contacts',
  POSTS = 'Posts',
  TODO = 'Todo',
  CREATE_REG_USER = 'Create and registration user',
  CREATE_CONTACT = 'Create contact',
  CREATE_TODO = 'Create task',
  ACCESS_TOKEN = 'accessToken',
  UPDATE_TOKEN = 'Refresh token',
  GET_CONTACTS_USER = 'Get contacts by user',
  GET_USERS = 'Get all users',
  GET_TODO = 'Get all tasks',
  GET_ONE_TODO = 'Get one task by id',
  UPDATE_TODO = 'Update task',
  REMOVE_TODO = 'Remove task by id',
  REMOVE_MARK_ID = 'Remove mark contact by id',
  REMOVE_M_CONTACTS = 'Remove all mark contacts',
  TOGGLE_MARK_REMOVE = 'Toggle mark for remove contacts',
  EDIT_CONTACT = 'Edit contact',
  BE_STRING = 'Should be a string',
  BE_NUMBER = 'Should be a number',
  BE_BOOL = 'Should be a boolean',
  TO_64 = 'Allowed to 64 characters',
  TO_128 = 'Allowed to 128 characters',
  TO_200 = 'Allowed to 200 characters',
  FROM_1_TO_64 = 'Allowed from 1 to 64 characters',
  FROM_2_TO_64 = 'Allowed from 2 to 64 characters',
  ASSIGN_ROLES = 'Assigning roles',
  BANNED_USER = 'Banned user',
}