// skills-server/src/core/helpers/endpoint.enum.ts
export enum EndpointEnum {
  ID = '/:id',
  PARAM_ID = 'id',
  MARK_ID = '/mark/:id',
  AUTH = 'auth',
  USERS = 'users',
  DOCS = '/docs',
  LOGIN = '/login',
  REGISTER = '/registration',
  REFRESH = '/refresh',
  ROLE = '/role',
  BAN = '/ban',
  VALUE = '/:value',
  EDIT = '/edit/:id',
  PARAM_VALUE = 'value',
  CONTACTS = 'contacts',
  POSTS = 'posts',
  IMAGE = 'image',
  ROLES = 'roles',
  TODO = 'todo',
  TASKS = 'tasks',
}